fix: image upload url test api
- move from feedback service to channel service
