+++ apps/api/src/domains/channel/channel/channel.service.ts
@@ -13,6 +13,8 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
+import { S3Client } from '@aws-sdk/client-s3';
+import { createPresignedPost } from '@aws-sdk/s3-presigned-post';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Transactional } from 'typeorm-transactional';
@@ -22,6 +24,7 @@ import { ProjectService } from '@/domains/project/project/project.service';
 import { FieldService } from '../field/field.service';
 import { ChannelMySQLService } from './channel.mysql.service';
 import type {
+  CreateImageUploadUrlDto,
   FindAllChannelsByProjectIdDto,
   FindByChannelIdDto,
   FindOneByNameAndProjectIdDto,
@@ -90,5 +93,32 @@ export class ChannelService {
     }
 
     await this.channelMySQLService.delete(channelId);
+  }
+
+  async createImageUploadUrl(dto: CreateImageUploadUrlDto) {
+    const {
+      projectId,
+      channelId,
+      accessKeyId,
+      secretAccessKey,
+      endpoint,
+      region,
+      bucket,
+    } = dto;
+
+    const s3 = new S3Client({
+      credentials: {
+        accessKeyId,
+        secretAccessKey,
+      },
+      endpoint,
+      region,
+    });
+
+    return await createPresignedPost(s3, {
+      Bucket: bucket,
+      Key: `${projectId}_${channelId}_${Date.now()}.png`,
+      Conditions: [{ 'Content-Type': 'image/png' }],
+    });
   }
 }
