+++ apps/api/src/domains/channel/channel/dtos/index.ts
@@ -19,3 +19,4 @@ export { UpdateChannelDto } from './update-channel.dto';
 export { FindByChannelIdDto } from './find-by-channel-id.dto';
 export { UpdateChannelFieldsDto } from './update-channel-fields.dto';
 export { FindOneByNameAndProjectIdDto } from './find-one-by-name-and-project-id.dto';
+export { CreateImageUploadUrlDto } from './create-image-upload-url.dto';
