+++ apps/api/src/domains/channel/channel/dtos/responses/index.ts
@@ -17,3 +17,4 @@ export { CreateChannelResponseDto } from './create-channel-response.dto';
 export { FindChannelByIdResponseDto } from './find-channel-by-id-response.dto';
 export { FindChannelsByProjectIdResponseDto } from './find-channels-by-id-response.dto';
 export { ImageConfigResponseDto } from './image-config-response.dto';
+export { ImageUploadUrlTestResponseDto } from './image-upload-url-test-response.dto';
