+++ apps/api/src/domains/feedback/feedback.controller.ts
@@ -42,12 +42,10 @@ import {
   DeleteFeedbacksRequestDto,
   ExportFeedbacksRequestDto,
   FindFeedbacksByChannelIdRequestDto,
-  ImageUploadUrlTestRequestDto,
 } from './dtos/requests';
 import {
   AddIssueResponseDto,
   FindFeedbacksByChannelIdResponseDto,
-  ImageUploadUrlTestResponseDto,
 } from './dtos/responses';
 import { FeedbackService } from './feedback.service';
 
@@ -192,39 +190,5 @@ export class FeedbackController {
     @Body() { feedbackIds }: DeleteFeedbacksRequestDto,
   ) {
     await this.feedbackService.deleteByIds({ channelId, feedbackIds });
-  }
-
-  @ApiParam({ name: 'projectId', type: Number })
-  @ApiOkResponse({ type: ImageUploadUrlTestResponseDto })
-  @Post('/image-upload-url-test')
-  async getImageUploadUrlTest(
-    @Body()
-    {
-      accessKeyId,
-      secretAccessKey,
-      endpoint,
-      region,
-      bucket,
-    }: ImageUploadUrlTestRequestDto,
-  ) {
-    try {
-      const presignedUrl = await this.feedbackService.createImageUploadUrl({
-        projectId: 0,
-        channelId: 0,
-        accessKeyId,
-        secretAccessKey,
-        endpoint,
-        region,
-        bucket,
-      });
-
-      if (presignedUrl.url) {
-        return { success: true };
-      } else {
-        return { success: false };
-      }
-    } catch (error) {
-      return { success: false };
-    }
   }
 }
